split behavior impl method validation
